Slightly improved usability

+++ datasets/head_phantom.sh
@@ -12,6 +12,7 @@ else
 fi
 
 # create python virtual environment
+python --version
 python -m venv .venv/
 source .venv/bin/activate
 
@@ -23,9 +24,14 @@ export PATH=$PATH:/home/runnerx/.local/bin
 # download head phantom data
 # first check if head phantom data folder exists already
 if [ -d "data" ]; then
-    echo "[INFO] data/ folder already exists - skipping download."
+    echo "[INFO] data/ folder found."
 else
-    echo "[INFO] Downloading head phantom maps (requires GitHub Actions secret)"
+    # Check if RDM_USER and RDM_KEY are set
+    if [ -z "$RDM_USER" ] || [ -z "$RDM_KEY" ]; then
+        echo "[ERROR] RDM_USER and/or RDM_KEY are not set! If you are running this on GitHub Actions, make sure you have set the secrets RDM_USER and RDM_KEY correctly. If you are running this locally, download the head phantom repository from https://doi.org/10.34973/m20r-jt17 and place the data/ directory here."
+        exit 1
+    fi
+    echo "[INFO] Downloading head phantom maps"
     python get-maps.py
     tar xf data.tar
     rm data.tar
@@ -34,5 +40,4 @@ fi
 # generate bids data
 echo "[INFO] Simulating BIDS dataset"
 qsm-forward head data/ bids
-rm -rf data/
